AcademiaVeritas/backend/requirements.txt
Flask>=2.0
Flask-Cors>=3.0
Flask-Dance[google]>=5.0
mysql-connector-python>=8.0
python-dotenv>=0.20
PyJWT>=2.0
Werkzeug>=2.0
gunicorn>=20.0
web3>=6.0
pytesseract>=0.3
Pillow>=9.0
pdf2image>=1.16
